Added previous change for country name validation and merge progress
Forgot to add the validation checking for the prototype as per the previous commit.

Current progress and findings for the dataset merge:

1. Dataset and array stacking not to be used due to lacking S5P metadata information compared to Sentinel 2 such as EPSG code, projection, resolution etc

2. Turns out the reason for difficulty in dataset merging is not due to data being lost but due to uneven shape of the dataset in the scanline direction which resulted in tons of null values.

Merging the datasets resulted in a shape with the longest dimension belonging to the dataset of the longest scanline. Other smaller scanline datasets have increased their scanline dimensions and filled it with NaN values